Allow skipping the installation step during setup
For integration testing, postgresql_embedded is really useful, as it allows us to delegate handling of local instance setup and teardown, makes creating sample dbs simple, etc. However, we don't generally like to pull random binaries off of the internet, and would prefer to provide our own postgres installation paths. This also makes it easier for things like bazel and nix which we use in our dev/CI environemnts.

This change just allows the setup() and new() methods to trust the installation dir that the user provided, and not try to guess an installation dir based on version, or attempt to install if the path doesn't match the version. The original behavior is still the default and how most people will use this crate.